+++ src/super/i-block/modules/state/index.ts
@@ -252,7 +252,8 @@ export default class State extends Friend {
 
 		const
 			{ctx} = this,
-			{router} = ctx.r;
+			{router} = ctx.r,
+			{routerStateUpdateMethod} = ctx;
 
 		data = ctx.syncRouterState(data, 'remote');
 		this.set(ctx.syncRouterState(data));
@@ -261,7 +262,7 @@ export default class State extends Friend {
 			return false;
 		}
 
-		await router.push(null, {
+		await router[routerStateUpdateMethod](null, {
 			query: data
 		});
 
@@ -388,7 +389,7 @@ export default class State extends Friend {
 	/**
 	 * Resets a router state of the current component
 	 */
-	async resetRouter(): Promise<boolean> {
+	resetRouter(): boolean {
 		//#if runtime has bRouter
 
 		const
@@ -406,7 +407,6 @@ export default class State extends Friend {
 			return false;
 		}
 
-		await router.push(null);
 		ctx.log('state:reset:router', this, stateFields);
 		return true;